+++ ClientApp/src/components/Account/AccountComp.jsx
@@ -4,47 +4,23 @@ import { Modal, ModalHeader, ModalBody, ModalFooter, Input, Label, NavLink } fro
 import { Button } from 'reactstrap';
 import { UserContext } from '../../UserContext';
 import { ConfirmationButton } from '../ConfirmationButton';
-import styled from "styled-components";
 import { AccountBox } from "./AccountCompLayout";
 
 export const AccountComp = () => {
+    
     const [modal, setModal] = useState(false);
-
-    const [username, setUsername] = useState('');
-    const [password, setPassword] = useState('');
-    const [incorrectUsernameOrPassword, setincorrectUsernameOrPassword] = useState(false);
 
     const { loggedIn, setLoggedIn, user: currentUser, setUser: setCurrentUser } = useContext(UserContext);
 
     const toggle = () => {
         setModal(!modal);
-        setincorrectUsernameOrPassword(false);
-    };
-
-    const onLogIn = async () => {
-        const response = await fetch('user/authenticate', {
-            method: 'POST',
-            headers: { 'Content-type': 'application/json' },
-            body: JSON.stringify({ username, password }),
-        });
-
-        if (response.ok) {
-            setincorrectUsernameOrPassword(false);
-            const user = JSON.parse(await response.text());
-            setCurrentUser(user);
-            setLoggedIn(true);
-            localStorage.setItem('token', user.token);
-            toggle();
-        }
-        else {
-            setincorrectUsernameOrPassword(true);
-        }
     };
 
     const onLogOut = () => {
         setLoggedIn(false);
         setCurrentUser({});
         localStorage.removeItem('token');
+        setModal(false);
     }
 
     return (loggedIn ?
@@ -53,8 +29,8 @@ export const AccountComp = () => {
                 {`${currentUser.firstName} ${currentUser.lastName}`}
             </NavLink>
             <ConfirmationButton onSubmit={onLogOut} submitLabel={'Log out'}
-                prompt={'Are you sure you want to log out?'}>Log out</ConfirmationButton>
-        </div>
+                prompt={'Are you sure you want to log out?'}>Log out</ConfirmationButton> 
+        </div> // #TODO open homepage when logging out
         : <div>
             <Button color="primary" onClick={toggle}>Log in</Button>
             <Modal className='Modal' autoFocus={false} isOpen={modal} toggle={toggle} size = "sm">
