+++ ClientApp/src/components/Account/LoginContext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const LoginContext = createContext(null);
